+++ src/backend/libinput/tablet.rs
@@ -103,12 +103,12 @@ where
         tablet_tool::TabletToolEventTrait::y(self)
     }
 
-    fn x_transformed(&self, width: u32) -> f64 {
-        tablet_tool::TabletToolEventTrait::x_transformed(self, width)
+    fn x_transformed(&self, width: i32) -> f64 {
+        tablet_tool::TabletToolEventTrait::x_transformed(self, width as u32)
     }
 
-    fn y_transformed(&self, height: u32) -> f64 {
-        tablet_tool::TabletToolEventTrait::y_transformed(self, height)
+    fn y_transformed(&self, height: i32) -> f64 {
+        tablet_tool::TabletToolEventTrait::y_transformed(self, height as u32)
     }
 
     fn distance(&self) -> f64 {
